requirements.txt
python-socketio[client]==5.11.4
eventlet==0.36.1
requests==2.32.3
